requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
psycopg2-binary
alembic
python-dotenv
pytest
pytest-asyncio
pytest-cov
requests